+++ prog/dftb+/lib_poisson/gallocation.F90
@@ -6,10 +6,12 @@
 !  under the LGPL licence.                                                *
 !**************************************************************************
 
+#:include "error.fypp"
+
 #!#:set MEMLOG = 1
 
 #! (LABEL, TYPE, ARRAY, ARGS) tuple for all logged arrays
-#:set ALLOC_CASES = [('l','logical',':', 'length'),&
+#:set ALLOC_CASES = [&
   & ('i', 'integer', ':', 'length'),&
   & ('d', 'real(dp)', ':', 'length'),&
   & ('z', 'complex(dp)', ':', 'length'),&
@@ -22,10 +24,10 @@
   & ('d4', 'real(dp)', ':,:,:,:', 'row,col,dep,qep')]
 
 module gallocation
-  use std_io
-  use dftbp_accuracy, only : dp, lc
-  use dftbp_message
   use, intrinsic :: iso_fortran_env, only : int64
+  use, intrinsic :: iso_c_binding, only : c_sizeof
+  use std_io
+  use dftbp_accuracy, only : dp
 
   integer, parameter :: long = int64
   integer, save :: iolog
@@ -51,30 +53,39 @@ contains
 #:for LABEL, TYPE, ARRAY, ARGS in ALLOC_CASES
 
   !---------------------------------------------------------------
-  subroutine allocate_${LABEL}$(array, ${ARGS}$)
+  subroutine allocate_${LABEL}$(array, ${ARGS}$, err)
 
-    ${TYPE}$, allocatable, intent(inout) :: array(${ARRAY}$)
+    ${TYPE}$, allocatable, target, intent(inout) :: array(${ARRAY}$)
 
     integer, intent(in) :: ${ARGS}$
 
+    !> Error code, 0 if no problems
+    integer, intent(out), optional :: err
+
+    ${TYPE}$, pointer :: pArrayFlat(:)
     integer :: iErr
-    character(lc) :: strTmp
+
+    if (present(err)) then
+      err = 0
+    end if
 
     !Allocation control: if array is already allocated STOP and write error statement
     if (allocated(array)) then
-      call error('ALLOCATION ERROR: ${TYPE}$ (${LABEL}$) array is already allocated')
+      @:ERROR_HANDLING(err,-1,"ALLOCATION ERROR: ${TYPE}$ (${LABEL}$) array is already allocated")
     endif
 
     if(.not. allocated(array)) then
       allocate(array(${ARGS}$), stat=iErr)
       if (ierr /= 0) then
-        write(strTmp, "(A,I0)")'Poisson allocation error: ', iErr
-        call error(strTmp)
+        @:FORMATTED_ERROR_HANDLING(err, iErr, "(A,I0)", "Poisson allocation error: ", iErr)
       else
-        alloc_mem= alloc_mem + sizeof(array)
-        if (alloc_mem > peak_mem) then
-          peak_mem = alloc_mem
-        endif
+        if (size(array) > 0) then
+          pArrayFlat(1 : size(array)) => array
+          alloc_mem= alloc_mem + c_sizeof(pArrayFlat(1)) * size(pArrayFlat)
+          if (alloc_mem > peak_mem) then
+            peak_mem = alloc_mem
+          endif
+        end if
       #:if defined('MEMLOG')
         call writePoissMemInfo()
       #:endif
@@ -89,18 +100,30 @@ contains
 #:for LABEL, TYPE, ARRAY, _ in ALLOC_CASES
 
   !---------------------------------------------------------------
-  subroutine deallocate_${LABEL}$(array)
+  subroutine deallocate_${LABEL}$(array, err)
 
-    ${TYPE}$, allocatable, intent(inout) :: array(${ARRAY}$)
+    ${TYPE}$, allocatable, target, intent(inout) :: array(${ARRAY}$)
+
+    !> Error code, 0 if no problems
+    integer, intent(out), optional :: err
+
+    ${TYPE}$, pointer :: pArrayFlat(:)
+
+    if (present(err)) then
+      err = 0
+    end if
 
     if (allocated(array)) then
-      alloc_mem= alloc_mem - sizeof(array)
+      if (size(array) > 0) then
+        pArrayFlat(1 : size(array)) => array
+        alloc_mem= alloc_mem - c_sizeof(pArrayFlat(1)) * size(pArrayFlat)
+      end if
       deallocate(array)
      #:if defined('MEMLOG')
       call writePoissMemInfo()
      #:endif
     else
-      call error('ALLOCATION ERROR: ${TYPE}$ (${LABEL}$) array is not allocated')
+      @:ERROR_HANDLING(err, -1, 'ALLOCATION ERROR: ${TYPE}$ (${LABEL}$) array is not allocated')
     endif
 
   end subroutine deallocate_${LABEL}$
@@ -154,13 +177,13 @@ contains
     real(dp), intent(out) :: dec
     character(3), intent(out) :: str
 
-    if (mem < 1000) then
+    if (mem < 1000_int64) then
       str = ' bt'
       dec = 1.0_dp
-    else if (mem < 1E7_int64) then
+    else if (mem < 10000000_int64) then
       str = ' kb'
       dec = 1.0E-3_dp
-    else if (mem < 1E10_int64) then
+    else if (mem < 10000000000_int64) then
       str = ' Mb'
       dec = 1.0E-6_dp
     else